bpo-40241: Add pycore_interp.h header
Move PyInterpreterState and related functions to a new internal
pycore_interp.h header file.
